feat(consensus): optimize cordial knowledge logic to reduce header synchronization requests (#9258)
# Description of change

PR removes the dissemination worker and makes updates in cordial
knowledge without additional channel messages.
We update cordial knowledge about the block ancestors when creating the
bundle, so the ancestors from useful headers authors are guaranteed to
be included in the bundle. It decreases the number of header
synchronisation requests.


## Links to any relevant issues

fixes #9214

## How the change has been tested

- [ ] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

---------

Co-authored-by: Nikita Polianskii <nikita.polianskii@iota.org>