feat: add retriers on create instance (#9297)
# Description of change

The `run_instances` aws command fails to return the minimum number of
instances for both spot and on-demand. It needs to be retried a few
times to collect the appropriate amount.

- requests spot instances in a loop for 5 minutes attempting to collect
as many available ones as needed.
- requests on-demand instances in a loop until the full quantity of
required instances is not collected.
- displays a status of the time left in the spot instance loop as well
as the number of instances collected.
- added the `assets/run_command.sh` script to execute a command on every
Node instance.
- adds `RUSTFLAGS=-Ctarget-cpu=native` build flag to optimize for CPU in
test available instructions.

## Links to any relevant issues

fixes #9274.

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes