feat(authority): make simulate_transaction work with both dry_run and dev_inspect (#9324)
# Description of change

Changed the implementation of `simulate_transaction` in the authority.
Now it supports the execution of both DRY RUN (when passing
`VmChecks::ENABLED`) and DEV INSPECT (when passing
`VmChecks::DISABLED`).

The REST API was not changed and a dry run was hardcoded for the
`simulate_transaction` API (as it was working before). Not sure if this
is needed to be changed as well. If yes, `TransactionSimulationResponse`
should be changed to include `execution_result` and
`SimulateTransactionQueryParameters` should be changed to include the
`VmChecks` param.

The current tests should check that the
`authority::simulate_transaction` dry run path works correctly. No new
tests were added for the dev inspect path. If we don't change the REST
API now, these new tests could be written once the gRPC is ready.

## Links to any relevant issues

Fixes #9323 

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes