feat(iota-types): Add scoring metrics type (#9176)
This PR adds a new `VersionedScoringMetrics` type. `ScoringMetricsV1` is
currently an atomic version of `MisbehaviorReportV1`, even though this
is expected to change in their future versions. `VersionedScoringMetrics
` should be used to store misbehavior metrics about the rest of the
committee and collected in the node, but also to store the information
received in the misbehavior reports received from other authorities.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

- [x] Protocol: Introduces the VersionedScoringMetrics type, used to
store misbehavior metrics about all authorities
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API: