feat(consensus): add traversed headers tracker (#9310)
# Description of change

PR adds traversed headers tracker in Linearizer. Now, when we are trying
to commit transactions, we are checking that the corresponding header
was actually traversed. It guarantees us that if for some block
transactions are committed, then the corresponding header is also
committed.

A protocol parameter was added for this change. It is enabled since
version 16 in all networks (but it affects only Starfish which is run
only on Devnet now).

## Links to any relevant issues

fixes #9302 

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes

### Release Notes

- [x] Protocol: logic for committing transactions in Starfish was
changed
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API: