fix(graphql-rpc): resolve module origin correctly in move abort errors (#9551)
# Description of change

This patch fixes an issue with the resolution of the errors of a failed
transactions in GraphQL. Namely, the resolver was assuming that the
location of the error was represented by the runtime id, instead of the
storage id.

This is not the case anymore: The storage id is used to represent the
move location where the error occurs.

After removing the obsolete logic, when we request the errors for the
failed transaction
```graphql
{
  transactionBlock(digest: "Facc1j5Y4u1uYuRXMYag83wGWSL1dYMLtHhkRCAYXRGQ") {
    effects {
      status
      errors
    }
  }
}
```

we get the following response

```json
{
  "data": {
    "transactionBlock": {
      "effects": {
        "status": "FAILURE",
        "errors": "Error in 6th command, from '0x9ce7f6e1575d59241a4e9d8029ed527ec01e867bfb8639698c483b5757bcd079::trading::execute_increase_order_internal' (instruction 91), abort code: 10"
      }
    }
  }
}

```
## Links to any relevant issues

Fixes #9539 

## How the change has been tested

Describe the tests that you ran to verify your changes.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)

### Infrastructure QA (only required for crates that are maintained by
@iotaledger/infrastructure)

- [ ] Synchronization of the indexer from genesis for a network
including migration objects.
- [ ] Restart of indexer synchronization locally without resetting the
database.
- [ ] Restart of indexer synchronization on a production-like database.
- [ ] Deployment of services using Docker.
- [ ] Verification of API backward compatibility.

THe patch does not affect any of these areas.

### Release Notes

- [x] GraphQL: Fixes resolution of move abort errors for failed
transactions