upstream(iota): move protocol config fetch to `--verify-compatibility` flag (#9653)
# Description of change

This PR fixes the issue for upgrade compatibility checks which appears
when trying to do an upgrade with an older CLI than the current protocol
version on the network, but without passing the `--verify-compatibility`
flag. The code is moved in the if branch as `--verify-compatibility` is
false by default.

## Links to any relevant issues

Upstream PR: https://github.com/MystenLabs/sui/pull/23433

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes

### Release Notes

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [x] CLI: Fixed a bug where an upgrade command would terminate early if
the CLI binary is not at the same protocol version or newer than the
network.
- [ ] Rust SDK:
- [ ] REST API: